refactor: group models
