+++ ios/MLCChat/Models/ExampleModelUrl.swift
@@ -1,0 +1,12 @@
+//
+//  ExampleModelUrl.swift
+//  MLCChat
+//
+
+import Foundation
+
+struct ExampleModelUrl: Identifiable {
+    let id = UUID()
+    let model_url: String
+    let local_id: String
+}
