+++ ios/MLCChat/Models/ParamsConfig.swift
@@ -1,0 +1,12 @@
+//
+//  ParamsConfig.swift
+//  MLCChat
+//
+
+struct ParamsConfig: Decodable {
+    struct ParamsRecord: Decodable {
+        let dataPath: String
+    }
+
+    let records: [ParamsRecord]
+}
